requirements.txt
Flask==3.0.2
ldap3==2.9.1
python-dotenv==1.0.1
gunicorn==21.2.0
pywinrm==0.4.3